Contract fixes for upgrade to newer compiler

+++ tokens/contracts/erc-721-minimal/ERC721BurnableUpgradeableMinimal.sol
@@ -28,7 +28,7 @@ abstract contract ERC721BurnableUpgradeableMinimal is Initializable, ContextUpgr
      */
     function burn(uint256 tokenId) public virtual {
         if(!_exists(tokenId)) {
-            address owner = address(tokenId >> 96);
+            address owner = address(uint160(tokenId >> 96));
             require(owner == _msgSender(), "ERC721Burnable: caller is not owner, not burn");
             _setBurned(tokenId);
         } else {
